fix(mempool): BIP-68 mempool version gate compares UNSIGNED (Core uint32_t parity)
checkSequenceLocksLocked short-circuited on a SIGNED `tx.Version < 2`, so a
high-bit version (int32 -2 == uint32 0xFFFFFFFE >= 2) skipped the mempool
sequence-lock check while block validation (chainmanager CalculateSequenceLocks)
enforces it — a mempool/consensus divergence. Cast to uint32 to match Core
(consensus/tx_verify.cpp:51) and the already-fixed CalculateSequenceLocks +
OP_CSV consensus gates (fd7115b).

Also corrects the stale W132-G13 P0-CDIV audit test. The consensus divergence
was fixed in fd7115b (2026-06-14): CalculateSequenceLocks + OP_CSV both compare
uint32(Version), exactly like Core. But the audit test (2026-05-17) still
asserted the pre-fix bug (MinHeight=-1, "want both -1 (the bug)") and therefore
failed once the code was corrected. It now asserts Core parity (MinHeight=104) —
what the production gate already computes — turning it into a proper regression
guard.

Context: Core has ALWAYS compared nVersion unsigned for BIP-68. Even as int32_t
nVersion, tx_verify.cpp used static_cast<uint32_t>(nVersion) >= 2 with the
comment "requires cast to unsigned otherwise ... half the range of nVersion
wouldn't support BIP 68". PR#29325 later changed storage to uint32_t (no
consensus change).

+++ internal/mempool/mempool.go
@@ -1780,7 +1780,12 @@ func (mp *Mempool) checkSequenceLocksLocked(tx *wire.MsgTx) error {
 		return nil
 	}
 	// BIP-68 only applies to v2+ transactions; spare the work otherwise.
-	if tx.Version < 2 {
+	// Compare UNSIGNED to match Core (consensus/tx_verify.cpp:51, uint32_t
+	// version): a high-bit version (int32 -2 == uint32 0xFFFFFFFE >= 2) still
+	// enforces BIP-68. A signed `< 2` here would skip the mempool sequence-lock
+	// check for such a tx while block validation (chainmanager) enforces it —
+	// mempool/consensus divergence. Mirrors CalculateSequenceLocks + OP_CSV.
+	if uint32(tx.Version) < 2 {
 		return nil
 	}
 	cs := mp.config.ChainState
